Significant breaking change.
We need to expose data urls to end clients, and there is some unpleasant
ambiguity contrasting URIs and DataURLs. This breaking change eliminates
the URI attribute, and unifies it with the DataURL. This does mean
making some changes to all fetchers, as DataURL now **must** be unique,
if necessary with synthetic fragments added.