Add Location constructor and function to truncate long/lat values.
This may be contentious, but in data reported by external IoT devices we
have some lat/long values with 13+ digits after the decimal place.  This
implies a precision of about 1 angstrom (i.e. about half the thickness
of an atom), which is obviously completely spurious.

Truncating these values to 6 decimal places gives a maximum precision of
~11 cm which I think is as accurate as we need.

This commit adds a `Truncate` function for truncating a float's decimal
places, and adds a `Location` constructor that truncates the given
values.

In addition we also add Longitude and Latitude type aliases of `float64`
to try and eliminate confusion when building these things.